+++ src/mako/application/web/Application.php
@@ -8,11 +8,15 @@
 namespace mako\application\web;
 
 use mako\application\Application as BaseApplication;
+use mako\application\DeferredTasks;
 use mako\http\Request;
 use mako\http\routing\Dispatcher;
 use mako\http\routing\Router;
 use mako\i18n\I18n;
+use mako\session\Session;
 
+use function fastcgi_finish_request;
+use function function_exists;
 use function ob_start;
 
 /**
@@ -20,6 +24,16 @@ use function ob_start;
  */
 class Application extends BaseApplication
 {
+	/**
+	 * {@inheritDoc}
+	 */
+	protected function initialize(): void
+	{
+		parent::initialize();
+
+		$this->container->registerSingleton(DeferredTasks::class, DeferredTasks::class);
+	}
+
 	/**
 	 * {@inheritDoc}
 	 */
@@ -46,5 +60,23 @@ class Application extends BaseApplication
 		// Dispatch the request and send the response
 
 		$this->container->get(Dispatcher::class)->dispatch($route)->send();
+
+		// Execute deferred tasks
+
+		if ($this->container->has(DeferredTasks::class)) {
+			if (function_exists('fastcgi_finish_request')) {
+				if ($this->container->has(Session::class)) {
+					$session = $this->container->get(Session::class);
+					$session->disableAutoCommit();
+					$session->commit();
+				}
+
+				fastcgi_finish_request();
+			}
+
+			foreach ($this->container->get(DeferredTasks::class)->getTasks() as $task) {
+				$this->container->call($task);
+			}
+		}
 	}
 }