+++ CHANGELOG.md
@@ -51,6 +51,7 @@ The major version bump is due to upping the required PHP version from `8.1` to `
 * Added a CLI spinner output helper.
 * Added a CLI hyperlink helper.
 * New and improved look of rendered CLI tables.
+* Added a simple way to defer small tasks until after the response has been sent to the client when using FastCGI.
 
 #### Changes